refactor: consolidate UI components
