+++ test/snackbar_test.dart
@@ -2,7 +2,7 @@ import 'package:flutter/material.dart';
 import 'package:flutter/services.dart';
 import 'package:flutter_test/flutter_test.dart';
 import 'package:notes_reminder_app/models/security_cue.dart';
-import 'package:notes_reminder_app/widgets/snackbar.dart';
+import 'package:notes_reminder_app/pandora_ui/snackbar.dart';
 
 void main() {
   TestWidgetsFlutterBinding.ensureInitialized();
