+++ test/teach_ai_modal_test.dart
@@ -2,7 +2,7 @@ import 'package:flutter/material.dart';
 import 'package:flutter/services.dart';
 import 'package:flutter_test/flutter_test.dart';
 import 'package:notes_reminder_app/models/security_cue.dart';
-import 'package:notes_reminder_app/widgets/teach_ai_modal.dart';
+import 'package:notes_reminder_app/pandora_ui/teach_ai_modal.dart';
 
 void main() {
   TestWidgetsFlutterBinding.ensureInitialized();
